halflife: keep game-DLL math single precision

+++ src/halflife/bmodels.cpp
@@ -42,7 +42,7 @@ extern DLL_GLOBAL Vector		g_vecAttackDir;
 //
 Vector VecBModelOrigin( entvars_t* pevBModel )
 {
-	return pevBModel->absmin + ( pevBModel->size * 0.5 );
+	return pevBModel->absmin + ( pevBModel->size * 0.5f );
 }
 
 // =================== FUNC_WALL ==============================================
@@ -179,7 +179,7 @@ void CFuncConveyor :: Spawn( void )
 void CFuncConveyor :: UpdateSpeed( float speed )
 {
 	// Encode it as an integer with 4 fractional bits
-	int speedCode = (int)(fabs(speed) * 16.0);
+	int speedCode = (int)(fabs(speed) * 16.0f);
 
 	if ( speed < 0 )
 		pev->rendercolor.x = 1;
@@ -339,17 +339,17 @@ void CFuncRotating :: KeyValue( KeyValueData* pkvd)
 {
 	if (FStrEq(pkvd->szKeyName, "fanfriction"))
 	{
-		m_flFanFriction = atof(pkvd->szValue)/100;
+		m_flFanFriction = (float)atof(pkvd->szValue) / 100.0f;
 		pkvd->fHandled = TRUE;
 	}
 	else if (FStrEq(pkvd->szKeyName, "Volume"))
 	{
-		m_flVolume = atof(pkvd->szValue)/10.0;
+		m_flVolume = (float)atof(pkvd->szValue) / 10.0f;
 
-		if (m_flVolume > 1.0)
-			m_flVolume = 1.0;
-		if (m_flVolume < 0.0)
-			m_flVolume = 0.0;
+		if (m_flVolume > 1.0f)
+			m_flVolume = 1.0f;
+		if (m_flVolume < 0.0f)
+			m_flVolume = 0.0f;
 		pkvd->fHandled = TRUE;
 	}
 	else if (FStrEq(pkvd->szKeyName, "spawnorigin"))
@@ -390,8 +390,8 @@ void CFuncRotating :: Spawn( )
 	m_pitch = PITCH_NORM - 1;
 
 	// maintain compatibility with previous maps
-	if (m_flVolume == 0.0)
-		m_flVolume = 1.0;
+	if (m_flVolume == 0.0f)
+		m_flVolume = 1.0f;
 
 	// if the designer didn't set a sound attenuation, default to one.
 	m_flAttenuation = ATTN_NORM;
@@ -455,7 +455,7 @@ void CFuncRotating :: Spawn( )
 	if ( FBitSet( pev->spawnflags, SF_BRUSH_ROTATE_INSTANT) )
 	{		
 		SetThink( SUB_CallUseToggle );
-		pev->nextthink = pev->ltime + 1.5;	// leave a magic delay for client to start up
+		pev->nextthink = pev->ltime + 1.5f;	// leave a magic delay for client to start up
 	}	
 	// can this brush inflict pain?
 	if ( FBitSet (pev->spawnflags, SF_BRUSH_HURT) )
@@ -528,7 +528,7 @@ void CFuncRotating :: Precache( void )
 		// make sure we restart the sound.  1.5 sec delay is magic number. KDB
 
 		SetThink ( SpinUp );
-		pev->nextthink = pev->ltime + 1.5;
+		pev->nextthink = pev->ltime + 1.5f;
 	}
 }
 
@@ -609,7 +609,7 @@ void CFuncRotating :: SpinUp( void )
 {
 	Vector	vecAVel;//rotational velocity
 
-	pev->nextthink = pev->ltime + 0.1;
+	pev->nextthink = pev->ltime + 0.1f;
 	pev->avelocity = pev->avelocity + ( pev->movedir * ( pev->speed * m_flFanFriction ) );
 
 	vecAVel = pev->avelocity;// cache entity's rotational velocity
@@ -640,7 +640,7 @@ void CFuncRotating :: SpinDown( void )
 	Vector	vecAVel;//rotational velocity
 	vec_t vecdir;
 
-	pev->nextthink = pev->ltime + 0.1;
+	pev->nextthink = pev->ltime + 0.1f;
 
 	pev->avelocity = pev->avelocity - ( pev->movedir * ( pev->speed * m_flFanFriction ) );//spin down slower than spinup
 
@@ -693,15 +693,15 @@ void CFuncRotating :: RotatingUse( CBaseEntity *pActivator, CBaseEntity *pCaller
 			//EMIT_SOUND_DYN(ENT(pev), CHAN_WEAPON, (char *)STRING(pev->noiseStop), 
 			//	m_flVolume, m_flAttenuation, 0, m_pitch);
 
-			pev->nextthink = pev->ltime + 0.1;
+			pev->nextthink = pev->ltime + 0.1f;
 		}
 		else// fan is not moving, so start it
 		{
 			SetThink ( SpinUp );
 			EMIT_SOUND_DYN(ENT(pev), CHAN_STATIC, (char *)STRING(pev->noiseRunning), 
-				0.01, m_flAttenuation, 0, FANPITCHMIN);
+				0.01f, m_flAttenuation, 0, FANPITCHMIN);
 
-			pev->nextthink = pev->ltime + 0.1;
+			pev->nextthink = pev->ltime + 0.1f;
 		}
 	}
 	else if ( !FBitSet ( pev->spawnflags, SF_BRUSH_ACCDCC ) )//this is a normal start/stop brush.
@@ -714,7 +714,7 @@ void CFuncRotating :: RotatingUse( CBaseEntity *pActivator, CBaseEntity *pCaller
 			// EMIT_SOUND_DYN(ENT(pev), CHAN_WEAPON, (char *)STRING(pev->noiseStop), 
 			//	m_flVolume, m_flAttenuation, 0, m_pitch);
 			
-			pev->nextthink = pev->ltime + 0.1;
+			pev->nextthink = pev->ltime + 0.1f;
 			// pev->avelocity = g_vecZero;
 		}
 		else
@@ -818,7 +818,7 @@ void CPendulum :: KeyValue( KeyValueData *pkvd )
 	}
 	else if (FStrEq(pkvd->szKeyName, "damp"))
 	{
-		m_damp = atof(pkvd->szValue) * 0.001;
+		m_damp = (float)atof(pkvd->szValue) * 0.001f;
 		pkvd->fHandled = TRUE;
 	}
 	else 
@@ -848,12 +848,12 @@ void CPendulum :: Spawn( void )
 	m_accel = (pev->speed * pev->speed) / (2 * fabs(m_distance));	// Calculate constant acceleration from speed and distance
 	m_maxSpeed = pev->speed;
 	m_start = pev->angles;
-	m_center = pev->angles + (m_distance * 0.5) * pev->movedir;
+	m_center = pev->angles + (m_distance * 0.5f) * pev->movedir;
 
 	if ( FBitSet( pev->spawnflags, SF_BRUSH_ROTATE_INSTANT) )
 	{		
 		SetThink( SUB_CallUseToggle );
-		pev->nextthink = gpGlobals->time + 0.1;
+		pev->nextthink = gpGlobals->time + 0.1f;
 	}
 	pev->speed = 0;
 	SetUse( PendulumUse );
@@ -888,7 +888,7 @@ void CPendulum :: PendulumUse( CBaseEntity *pActivator, CBaseEntity *pCaller, US
 	}
 	else
 	{
-		pev->nextthink = pev->ltime + 0.1;		// Start the pendulum moving
+		pev->nextthink = pev->ltime + 0.1f;		// Start the pendulum moving
 		m_time = gpGlobals->time;		// Save time to calculate dt
 		SetThink( Swing );
 		m_dampSpeed = m_maxSpeed;
@@ -932,12 +932,12 @@ void CPendulum :: Swing( void )
 	pev->avelocity = pev->speed * pev->movedir;
 
 	// Call this again
-	pev->nextthink = pev->ltime + 0.1;
+	pev->nextthink = pev->ltime + 0.1f;
 
 	if ( m_damp )
 	{
 		m_dampSpeed -= m_damp * m_dampSpeed * dt;
-		if ( m_dampSpeed < 30.0 )
+		if ( m_dampSpeed < 30.0f )
 		{
 			pev->angles = m_center;
 			pev->speed = 0;
@@ -965,7 +965,7 @@ void CPendulum :: Touch ( CBaseEntity *pOther )
 		return;
 
 	// calculate damage based on rotation speed
-	float damage = pev->dmg * pev->speed * 0.01;
+	float damage = pev->dmg * pev->speed * 0.01f;
 
 	if ( damage < 0 )
 		damage = -damage;
